+++ docs-beta/src/data/FeatureList.ts
@@ -1,0 +1,14 @@
+export const FeatureList=[
+    {
+      title: 'Simple pipeline definition',
+      Svg: require('../../static/img/icon-clock-time.svg').default,
+    },
+    {
+      title: 'Speedy local incremental builds',
+      Svg: require('../../static/img/icon-up.svg').default,
+    },
+    {
+      title: 'Scoped task runs',
+      Svg: require('../../static/img/icon-hierarchy.svg').default,
+    },
+  ];
