+++ docs-beta/src/data/ToolList.ts
@@ -1,22 +1,22 @@
 export const ToolList=[
     {
       title: 'Docusaurus',
-      Svg: require('../../../static/img/icon-clock-time.svg').default,
+      Svg: require('../../static/img/icon-clock-time.svg').default,
       description: "This is what it does. We were inspired to create something even better by using them.",
     },
     {
       title: 'Beachball',
-      Svg: require('../../../static/img/icon-up.svg').default,
+      Svg: require('../../static/img/icon-up.svg').default,
       description: "This is what it does. We were inspired to create something even better by using them.",
     },
     {
       title: 'Backfill',
-      Svg: require('../../../static/img/icon-hierarchy.svg').default,
+      Svg: require('../../static/img/icon-hierarchy.svg').default,
       description: "This is what it does. We were inspired to create something even better by using them.",
     },
     {
       title: 'P-graph',
-      Svg: require('../../../static/img/icon-hierarchy.svg').default,
+      Svg: require('../../static/img/icon-hierarchy.svg').default,
       description: "This is what it does. We were inspired to create something even better by using them.",
     },
   ];
