--- docs-beta/src/pages/data/TableContents.ts
@@ -1,156 +1,0 @@
-const lage = {
-  "name": "Lage",
-  "capabilities": {
-     "Local computation caching": "Yes", 
-     "Local task orchastration": "Yes",
-     "Distributed computation caching": "Yes", 
-     "Distributed task execution": "Yes",
-     "Transparent remote execution": "Yes", 
-     "Detecting affected projects/packages": "Yes",
-     "Workspace analysis": "Yes", 
-     "Dependency graph visualization": "Yes",
-     "Source code sharing": "Yes", 
-     "Consistent tooling": "Yes",
-     "Code generation": "Yes", 
-     "Project constraints and visibility": "Yes"
-    }
- };
-
-const bazel = {
-  "name": "Bazel",
-  "capabilities": {
-     "Local computation caching": "Yes", 
-     "Local task orchastration": "Yes",
-     "Distributed computation caching": "Yes", 
-     "Distributed task execution": "Yes",
-     "Transparent remote execution": "Yes", 
-     "Detecting affected projects/packages": "Yes",
-     "Workspace analysis": "Yes", 
-     "Dependency graph visualization": "Yes",
-     "Source code sharing": "Yes", 
-     "Consistent tooling": "Yes",
-     "Code generation": "Yes", 
-     "Project constraints and visibility": "Yes"
-    }
- };
-
-const gradle = {
-  "name": "Gradle",
-  "capabilities": {
-     "Local computation caching": "Yes", 
-     "Local task orchastration": "Yes",
-     "Distributed computation caching": "Yes", 
-     "Distributed task execution": "Yes",
-     "Transparent remote execution": "Yes", 
-     "Detecting affected projects/packages": "Yes",
-     "Workspace analysis": "Yes", 
-     "Dependency graph visualization": "Yes",
-     "Source code sharing": "Yes", 
-     "Consistent tooling": "Yes",
-     "Code generation": "Yes", 
-     "Project constraints and visibility": "Yes"
-    }
- };
-
-const nx = {
-  "name": "Nx",
-  "capabilities": {
-     "Local computation caching": "Yes", 
-     "Local task orchastration": "Yes",
-     "Distributed computation caching": "Yes", 
-     "Distributed task execution": "Yes",
-     "Transparent remote execution": "Yes", 
-     "Detecting affected projects/packages": "Yes",
-     "Workspace analysis": "Yes", 
-     "Dependency graph visualization": "Yes",
-     "Source code sharing": "Yes", 
-     "Consistent tooling": "Yes",
-     "Code generation": "Yes", 
-     "Project constraints and visibility": "Yes"
-    }
- };
-
-const lerna = {
-  "name": "Lerna",
-  "capabilities": {
-     "Local computation caching": "Yes", 
-     "Local task orchastration": "Yes",
-     "Distributed computation caching": "Yes", 
-     "Distributed task execution": "Yes",
-     "Transparent remote execution": "Yes", 
-     "Detecting affected projects/packages": "Yes",
-     "Workspace analysis": "Yes", 
-     "Dependency graph visualization": "Yes",
-     "Source code sharing": "Yes", 
-     "Consistent tooling": "Yes",
-     "Code generation": "Yes", 
-     "Project constraints and visibility": "Yes"
-    }
- };
-
-const pants = {
-  "name": "Pants",
-  "capabilities": {
-     "Local computation caching": "Yes", 
-     "Local task orchastration": "Yes",
-     "Distributed computation caching": "Yes", 
-     "Distributed task execution": "Yes",
-     "Transparent remote execution": "Yes", 
-     "Detecting affected projects/packages": "Yes",
-     "Workspace analysis": "Yes", 
-     "Dependency graph visualization": "Yes",
-     "Source code sharing": "Yes", 
-     "Consistent tooling": "Yes",
-     "Code generation": "Yes", 
-     "Project constraints and visibility": "Yes"
-    }
- };
-
-const rush = {
-  "name": "Rush",
-  "capabilities": {
-     "Local computation caching": "Yes", 
-     "Local task orchastration": "Yes",
-     "Distributed computation caching": "Yes", 
-     "Distributed task execution": "Yes",
-     "Transparent remote execution": "Yes", 
-     "Detecting affected projects/packages": "Yes",
-     "Workspace analysis": "Yes", 
-     "Dependency graph visualization": "Yes",
-     "Source code sharing": "Yes", 
-     "Consistent tooling": "Yes",
-     "Code generation": "Yes", 
-     "Project constraints and visibility": "Yes"
-    }
- };
-
-const turborepo = {
-  "name": "Turborepo",
-  "capabilities": {
-     "Local computation caching": "Yes", 
-     "Local task orchastration": "Yes",
-     "Distributed computation caching": "Yes", 
-     "Distributed task execution": "Yes",
-     "Transparent remote execution": "Yes", 
-     "Detecting affected projects/packages": "Yes",
-     "Workspace analysis": "Yes", 
-     "Dependency graph visualization": "Yes",
-     "Source code sharing": "Yes", 
-     "Consistent tooling": "Yes",
-     "Code generation": "Yes", 
-     "Project constraints and visibility": "Yes"
-    }
- };
-
-export const tableContents = [
-      lage,
-      bazel,
-      gradle,
-      nx,
-      lerna,
-      pants,
-      rush,
-      turborepo
-    ]
-  
-
