+++ package.json
@@ -24,6 +24,6 @@
     "beachball": "^2.16.0",
     "gh-pages": "^2.2.0",
     "jest": "^27.2.0",
-    "lage-npm": "npm:lage@^1.5.0"
+    "lage-npm": "npm:lage@^1.6.2"
   }
 }
